perf: optimize serialization perf (#3656)
## Why?

Improve serialization/deserialization throughput across Fory runtimes
and keep the cross-language benchmark suite current, comparable, and
documented.

## What does this PR do?

- Optimizes hot serialization/deserialization paths in
JavaScript/TypeScript, Go, C++, C#, Python, and Rust, including ref
reset skips, root serializer/deserializer caches, TypeMeta/read-schema
caches, string and varint fast paths, small typed-array writes, Go
varint32 struct reads, C++ root struct type info caching, C# inline
TypeMeta indexes, Python declared ndarray fields, and Rust signed varint
encoding.
- Adds Java xlang JMH benchmarks with FlatBuffers coverage, generated
FlatBuffers sources, a benchmark runner, and a report renderer.
- Refreshes benchmark payloads by renaming the simple struct fixture to
`NumericStruct`, expanding it to 12 int32 fields, and aligning
protobuf/msgpack/json benchmark coverage across languages.
- Updates benchmark runners and report generators for C++, C#, Dart, Go,
Java, JavaScript, Python, Rust, and Swift, including combined throughput
plots, serialized-size output, and regenerated `docs/benchmarks/**`
artifacts.
- Adds/updates focused coverage for the changed paths, including
JavaScript TypeMeta-compatible reader tests and Dart generated benchmark
model round-trip tests.

## Related issues

#1017 


## AI Contribution Checklist



- [ ] Substantial AI assistance was used in this PR: `yes` / `no`
- [ ] If `yes`, I included a completed [AI Contribution
Checklist](https://github.com/apache/fory/blob/main/AI_POLICY.md#9-contributor-checklist-for-ai-assisted-prs)
in this PR description and the required `AI Usage Disclosure`.
- [ ] If `yes`, my PR description includes the required `ai_review`
summary and screenshot evidence of the final clean AI review results
from both fresh reviewers on the current PR diff or current HEAD after
the latest code changes.



## Does this PR introduce any user-facing change?



- [ ] Does this PR introduce any public API change?
- [ ] Does this PR introduce any binary protocol compatibility change?

## Benchmark

- Refreshes benchmark reports and images under `docs/benchmarks/**` for
C++, C#, Dart, Go, Java, JavaScript, Python, Rust, and Swift.
- Adds Java xlang throughput reporting; the generated Java report
documents JMH parameters: `-f 1 -wi 3 -i 3 -t 1 -w 3s -r 3s -bm thrpt
-tu s`.

+++ rust/fory-core/src/buffer.rs
@@ -115,8 +115,8 @@ impl<'a> Writer<'a> {
 
     #[inline(always)]
     pub fn write_var_i32(&mut self, value: i32) {
-        let zigzag = ((value as i64) << 1) ^ ((value as i64) >> 31);
-        self._write_var_u32(zigzag as u32)
+        let zigzag = ((value as u32) << 1) ^ ((value >> 31) as u32);
+        self._write_var_u32(zigzag)
     }
 
     // ============ INT64 (TypeId = 6) ============
@@ -130,7 +130,7 @@ impl<'a> Writer<'a> {
 
     #[inline(always)]
     pub fn write_var_i64(&mut self, value: i64) {
-        let zigzag = ((value << 1) ^ (value >> 63)) as u64;
+        let zigzag = ((value as u64) << 1) ^ ((value >> 63) as u64);
         self._write_var_u64(zigzag);
     }
